添加优惠券管理接口

+++ mall-mbg/src/main/java/com/macro/mall/model/SmsCouponProductRelation.java
@@ -5,7 +5,7 @@ import java.io.Serializable;
 public class SmsCouponProductRelation implements Serializable {
     private Long id;
 
-    private Integer couponId;
+    private Long couponId;
 
     private Long productId;
 
@@ -19,11 +19,11 @@ public class SmsCouponProductRelation implements Serializable {
         this.id = id;
     }
 
-    public Integer getCouponId() {
+    public Long getCouponId() {
         return couponId;
     }
 
-    public void setCouponId(Integer couponId) {
+    public void setCouponId(Long couponId) {
         this.couponId = couponId;
     }
